refactor(api): reorganize for versioning
