+++ pallets/api-vnext/src/fungibles/precompiles/interfaces/v0/IFungibles.sol
