+++ pop-api-vnext/src/fungibles/v0/errors.rs
